% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{add3}
\alias{add3}
\title{Prepare User input.}
\usage{
add3(x)
}
\arguments{
\item{x}{A number}
}
\value{
Returns a number.
}
\description{
The response from GPT3 sometimes contains strings that are not R commands.
}
