% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{allcounts4su}
\alias{allcounts4su}
\title{A list containing total and 4su counts for exons and introns, three replicates and 11 time points: 0,1/6,1/3,1/2,1,1.5,2,4,8,12,16 hours.}
\format{An INSPEcT object}
\description{
A list containing total and 4su counts for exons and introns, three replicates and 11 time points: 0,1/6,1/3,1/2,1,1.5,2,4,8,12,16 hours.
}
