% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ratePvals-methods.R
\docType{methods}
\name{calculateRatePvals}
\alias{calculateRatePvals}
\alias{calculateRatePvals,INSPEcT-method}
\title{Calculate a single p-value for each rate}
\usage{
calculateRatePvals(object, bTsh = NULL, cTsh = NULL)

\S4method{calculateRatePvals}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT or INSPEcT_model}
}
\value{
A matrix containing p-values calculated for each rate
}
\description{
This method is used to calculate all the p-values relative to the variability of synthesis, processing and degradation rates.
In case the 'llr' method is chose via \code{\link{modelSelection}}, combined with Brown's method that combines the results of the 
log likelihood ratio test results for all pairs tested for each rate and all genes. P-values will change 
according to the threshold set for the chi-squared test because it influences the model that will be taken into 
consideration to perform log likelihood ratio tests. To have a sense of the best parameter to choose, a sythetic
data-set can be built and tested (\code{\link{makeSimModel}}, \code{\link{makeSimDataset}})
In case 'aic' has been selected via \code{\link{modelSelection}} method, 
this method assigns the chi-squared test result of the model selected by AIC
to the respective variable rates
}
\details{
ratePvlas retrieve a single p-value for each rate thanks to multiple log likelihood tests performed on 
nested models that has a chi-squared test below the selected threshold. 
Among the many p-values that log likelihood ratio test calculate, a single p-value is obtaied applying Brown's method 
for combining dependent p-values.
}
\examples{
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
calculateRatePvals(nascentInspObj10)
# Set the chi-squared threshold at .2 for nascentInspObj10 object
modelSelection(nascentInspObj10)$thresholds$chisquare <- .2
calculateRatePvals(nascentInspObj10)
}
\seealso{
\code{\link{makeSimModel}}, \code{\link{makeSimDataset}}
}
