% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/correlationPlot-methods.R
\name{correlationPlot}
\alias{correlationPlot}
\alias{correlationPlot,INSPEcT_model,INSPEcT-method}
\title{Display rate classification performance}
\usage{
correlationPlot(object, object2, plot = TRUE)

\S4method{correlationPlot}{INSPEcT_model,INSPEcT}(object, object2, plot = TRUE)
}
\arguments{
\item{object}{An object of class INSPEcT_model with simulated rates.}

\item{object2}{An object of class INSPEcT.}

\item{plot}{A logical indicating whether to draw or not the plot. (default=TRUE)}
}
\value{
An list with the correlation values.
}
\description{
This function plot the rates of a simulated dataset against the modeled ones and compute their correlations.
}
