% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{dim,INSPEcT-method}
\alias{dim,INSPEcT-method}
\title{Dimensions of an Object of Class INSPEcT}
\usage{
\S4method{dim}{INSPEcT}(x)
}
\arguments{
\item{x}{An object of class INSPEcT}
}
\value{
A numeric that indicates the number of genes within the object and 
the number of time points contained the object
}
\description{
A method to obtain the dimension of the object of class INSPEcT reported as a vector
containing of the genes and the number of time points
}
\seealso{
\code{\link{nGenes}}, \code{\link{nTpts}}
}

