% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{eda}
\alias{eda}
\title{Create a report for exploratory data analysis (EDA)}
\usage{
eda(df, target = NULL)
}
\arguments{
\item{df}{A data.frame or a matrix with at least 2 columns.}

\item{target}{A selected target variable.}
}
\value{
This function does not return a value. Instead, it generates an HTML file containing the EDA report and attempts to open this file in the default web browser.
}
\description{
This function generates a comprehensive EDA report and attempts to open it in a browser.
}
\details{
The function performs a check on the input data to ensure it is either a data frame or matrix.
It also checks that the input data has at least two columns. The function uses a predefined Rmarkdown template (`eda.Rmd`)
to generate the report. This template should be present in the `inspect` package.
}
\examples{
# Example with default parameters (using iris dataset)
eda(iris)

# Example with a target variable
eda(iris, "Species")

# Example with custom data frame and target variable
my_data <- data.frame(x = rnorm(100), y = rnorm(100))
eda(df = my_data, target = "x")

}
\seealso{
\code{\link[rmarkdown]{render}} for details on the rendering of Rmarkdown documents.
\code{\link[utils]{browseURL}} for how URLs or files are opened in browsers.
}
