% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{featureNames,INSPEcT-method}
\alias{featureNames,INSPEcT-method}
\alias{featureNames<-,INSPEcT-method}
\title{Gene Names Associated with an Object of Class INSPEcT}
\usage{
\S4method{featureNames}{INSPEcT}(object)

\S4method{featureNames}{INSPEcT}(object) <- value
}
\arguments{
\item{object}{An object of class INSPEcT}

\item{value}{A character that will replace the current feature names}
}
\value{
A character that contains gene names associated with the object of class INSPEcT
}
\description{
A method to visualize gene names associated with the object of class INSPEcT
}

