% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{featureWidths}
\alias{featureWidths}
\title{Contains two variables: "exWdths" and "intWdths" containing the lenght of the exons and introns, 
respectively, relative to the genes in "allcounts"}
\format{numeric vector of length 500}
\description{
Contains two variables: "exWdths" and "intWdths" containing the lenght of the exons and introns, 
respectively, relative to the genes in "allcounts"
}
