% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getModel-methods.R
\docType{methods}
\name{getModel}
\alias{getModel}
\alias{getModel<-}
\alias{getModel,INSPEcT-method}
\alias{getModel<-,INSPEcT-method}
\title{Get or replace INSPEcT_model object within INSPEcT object}
\usage{
getModel(object)

getModel(object) <- value

\S4method{getModel}{INSPEcT}(object)

\S4method{getModel}{INSPEcT}(object) <- value
}
\arguments{
\item{object}{An object of class INSPEcT}

\item{value}{An object of class INSPEcT model}
}
\value{
An object of class INSPEcT model
}
\description{
A method to get or set the INSPEcT_model object within an INSPEcT object. This method
is particularly useful to get and set testing parameters of the INSPEcT_model object
within the INSPEcT object.
}
\examples{
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
getModel(nascentInspObj10)
}
\seealso{
\code{\link{testingParams}}
}
