% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApps.R
\docType{methods}
\name{INSPEcT-shinyApps}
\alias{INSPEcT-shinyApps}
\alias{runProcessingRateDelay}
\alias{runINSPEcTGUI}
\title{Run shiny applications contained in the package INSPEcT}
\usage{
runProcessingRateDelay()

runINSPEcTGUI()
}
\description{
Two shiny apps are encoded into the package inspect:
- runProcessingRateDelay: plots single genes as well as genome wide plots associated to the
  processing induced delay loading the data from an INSPEcT object.
- runINSPEcTGUI: is a way to visualize and interact with the RNAdynamics at the level of 
  a single gene, either loading the data from ad INSPEcT object or from scratch.
}
\examples{
# runProcessingRateDelay()
# runINSPEcTGUI()
}
