% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{package}
\name{INSPEcT}
\alias{INSPEcT}
\title{INSPEcT package}
\description{
INSPEcT (INference of Synthesis, Processing and dEgradation rates from Transcriptomic data),
is a package that analyse RNA-seq data in order to evaluate synthesis, processing and degradation
rates and asses via modeling the rates that determines changes in RNA levels.

To see how the typical workflow of INSPEcT works, type: \cr
\code{vignette('INSPEcT')}

INSPEcT implements two main classes (\code{\linkS4class{INSPEcT}} and 
\code{\linkS4class{INSPEcT_diffsteady}}) and their corresponding methods. 
To have a detailed description of how the two classes are structured 
and which methods apply on, type:

\code{?'INSPEcT-class'} \cr
\code{?'INSPEcT_diffsteady-class'}


To obtain the citation, type: \cr
\code{citation('INSPEcT')}
}
