% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_bf}
\alias{inspect_bf}
\title{Check if an object is a numeric vector of valid Bayes factor values}
\usage{
inspect_bf(x)
}
\arguments{
\item{x}{An arbitrary object.}
}
\value{
\code{inspect_bf} does not return any output. There are three possible outcomes:
\itemize{
\item The call is silent if \code{x} is a numeric vector of valid Bayes factor values and there are no \code{NA} or \code{NaN} values.
\item An informative warning message is given if \code{p} is a numeric vector of valid Bayes factor values and there are \code{NA} or \code{NaN} values.
\item An informative error message is thrown if \code{p} is not a numeric vector of valid Bayes factor values. This will \code{\link[base]{stop}} the execution.
}
}
\description{
\code{inspect_bf} checks if an object is a numeric vector of valid Bayes factor values. This can be useful to validate inputs, intermediate calculations or outputs in user-defined functions.
}
\details{
\code{inspect_bf} conducts a series of tests to check if \code{x} is a numeric vector of valid Bayes factor values. Namely, \code{inspect_bf} checks if:
\itemize{
\item \code{x} is \code{NULL} or empty.
\item \code{x} is a numeric, atomic vector.
\item \code{x} has \code{NA} or \code{NaN} values.
\item The values of \code{x} are non-negative.
}
}
\examples{
# Calls that pass silently:
x <- c(0, 0.5, 1, 10, 50, 100)
inspect_bf(x)
inspect_bf(seq(10, 100, 10))

# Call that throws an informative warning message:
\dontrun{y <- c(0.1, 0.2, NA, 0.4, 0.5)}
\dontrun{inspect_bf(y)}

# Calls that throw informative error messages:
\dontrun{z <- c(-0.9, 0, 0.1, 0.2, 0.3, 0.4, 0.5)}
\dontrun{inspect_bf(z)}
\dontrun{mylist <- list(NULL, TRUE, factor(.5), matrix(0.5),
         "0.5", list(0.5), NA, NaN, -0.5, -5)}
\dontrun{inspect_bf(mylist)}
\dontrun{inspect_bf(mylist[[1]])}
\dontrun{inspect_bf(mylist[[2]])}
\dontrun{inspect_bf(mylist[[3]])}
\dontrun{inspect_bf(mylist[[4]])}
\dontrun{inspect_bf(mylist[[5]])}
\dontrun{inspect_bf(mylist[[6]])}
\dontrun{inspect_bf(mylist[[7]])}
\dontrun{inspect_bf(mylist[[8]])}
\dontrun{inspect_bf(mylist[[9]])}
\dontrun{inspect_bf(mylist[[10]])}

}
\seealso{
\itemize{
\item \code{\link[inspect]{inspect_log_bf}} to check if an object is a numeric vector of valid logarithmic Bayes factor values.
\item \code{\link[inspect]{inspect_prob}} to check if an object is a numeric vector of valid probability values.
\item \code{\link[inspect]{inspect_log_base}} to check if an object is a numeric vector of \code{\link[base]{length}} 1 representing a valid logarithmic base.
\item \code{\link[inspect]{inspect_scale}} to check if an object is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package.
}
}
