% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_character}
\alias{inspect_character}
\title{lorem ipsum}
\usage{
inspect_character(x, allowed)
}
\arguments{
\item{x}{lorem ipsumlorem ipsum}
}
\value{
lorem ipsum
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
