% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data_.R
\name{inspect_data_multinom_as_bern}
\alias{inspect_data_multinom_as_bern}
\title{lorem ipsum}
\usage{
inspect_data_multinom_as_bern(data, success, warning_nas = FALSE)
}
\arguments{
\item{x}{lorem ipsumlorem ipsum}
}
\value{
lorem ipsum
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
