% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data_.R
\name{inspect_data_multinomial}
\alias{inspect_data_multinomial}
\title{Validate categorical data}
\usage{
inspect_data_multinomial(data, warning_nas = FALSE)
}
\description{
\code{inspect_data_multinomial} checks if an object is a numeric vector of data eligible to have been generated by a Multinomial distribution. This can be useful to validate inputs in user-defined functions.
}
