% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_log_bfactor}
\alias{inspect_log_bfactor}
\title{Check if an object is a numeric vector of valid logarithmic Bayes factor values}
\usage{
inspect_log_bfactor(x, allow_nas = TRUE, warning_nas = TRUE)
}
\arguments{
\item{x}{An arbitrary object.}

\item{allow_nas}{Logical value. If \code{TRUE} then \code{NA} and \code{NaN} values in \code{x} are allowed. If \code{FALSE}, execution is stopped and an error message is thrown in case there are \code{NA} or \code{NaN} values in \code{x}.}

\item{warning_nas}{Logical value. If \code{TRUE} then the presence of \code{NA} or \code{NaN} values in \code{x} generates a warning message. \code{NA} and \code{NaN} pass silently otherwise (if \code{allow_nas} is \code{TRUE}).}
}
\value{
\code{inspect_log_bfactor} does not return any output. There are three possible outcomes:
\itemize{
\item The call is silent if:
\itemize{
\item \code{x} is a numeric vector of valid logarithmic Bayes factor values without \code{NA} or \code{NaN} values.
\item \code{x} is a numeric vector of valid logarithmic Bayes factor values with some \code{NA} or \code{NaN} values, \code{allow_nas} is set to \code{TRUE} and \code{warning_nas} is set to \code{FALSE}.
}
\item An informative warning message is given if \code{x} is a numeric vector of valid logarithmic Bayes factor values, there are \code{NA} or \code{NaN} values in \code{x} and both  \code{allow_nas} and \code{warning_nas} are set to \code{TRUE}.
\item An informative error message is thrown and the execution is stopped if:
\itemize{
\item \code{x} is not a numeric vector of valid logarithmic Bayes factor values.
\item \code{x} is a numeric vector of valid logarithmic Bayes factor values with some \code{NA} or \code{NaN} values and \code{allow_nas} is set to \code{FALSE}.
}
}
}
\description{
\code{inspect_log_bfactor} checks if an object is a numeric vector of valid logarithmic Bayes factor values.  This can be useful to validate inputs, intermediate calculations or outputs in user-defined functions.
}
\details{
\code{inspect_log_bfactor} conducts a series of tests to check if \code{x} is a numeric vector of valid logarithmic Bayes factor values. Namely, \code{inspect_log_bfactor} checks if:
\itemize{
\item \code{x} is \code{NULL} or empty.
\item \code{x} is a numeric, atomic vector.
\item \code{x} has \code{NA} or \code{NaN} values.
}
}
\examples{
# Calls that pass silently:
x1 <- c(0, 0.5, 1, 10, 50, 100)
x2 <- c(NA, 0.5, 1, 10, 50, 100)
inspect_log_bfactor(x1)
inspect_log_bfactor(x2, warning_nas = FALSE)
inspect_log_bfactor(x2, allow_nas = TRUE, warning_nas = FALSE)

# Call that throws an informative warning message:
\dontrun{y <- c(0.1, 0.2, NA, 0.4, 0.5)}
\dontrun{inspect_log_bfactor(y)}
\dontrun{inspect_log_bfactor(y, warning_nas = TRUE)}
\dontrun{inspect_log_bfactor(y, allow_nas = TRUE, warning_nas = TRUE)}

# Calls that throw informative error messages:
\dontrun{mylist <- list(NULL, TRUE, factor(.5), matrix(0.5),
         "0.5", list(0.5), numeric(0), NA, NaN)}
\dontrun{inspect_log_bfactor(mylist[[1]])}
\dontrun{inspect_log_bfactor(mylist[[2]])}
\dontrun{inspect_log_bfactor(mylist[[3]])}
\dontrun{inspect_log_bfactor(mylist[[4]])}
\dontrun{inspect_log_bfactor(mylist[[5]])}
\dontrun{inspect_log_bfactor(mylist[[6]])}
\dontrun{inspect_log_bfactor(mylist[[7]])}
\dontrun{inspect_log_bfactor(mylist[[8]])}
\dontrun{inspect_log_bfactor(mylist[[9]])}

}
\seealso{
\itemize{
\item \code{\link[inspector]{inspect_bfactor}} to check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[inspector]{inspect_prob}} to check if an object is a numeric vector of valid probability values.
}
}
