% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_parameter_.R
\name{inspect_par_haldane}
\alias{inspect_par_haldane}
\title{lorem ipsum}
\usage{
inspect_par_haldane(x)
}
\arguments{
\item{x}{lorem ipsumlorem ipsum}
}
\value{
lorem ipsum
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
