% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_prob}
\alias{inspect_prob}
\title{Check if an object is a vector of valid probability values}
\usage{
inspect_prob(x, allow_nas = TRUE)
}
\arguments{
\item{x}{An arbitrary object.}

\item{allow_nas}{Logical value indicating either \code{TRUE} or \code{FALSE}. If \code{FALSE} then the execution is stopped (via \code{\link[base]{stop}}) in case there are \code{NA} or \code{NaN} values in \code{x}.}
}
\value{
\code{inspect_prob} does not return any output. There are three possible outcomes:
\itemize{
\item The call is silent if \code{x} is a numeric vector of valid probability values without \code{NA} or \code{NaN} values.
\item An informative warning message is thrown if \code{x} is a numeric vector of valid probability values with some \code{NA} or \code{NaN} values and \code{allow_nas} is set to \code{TRUE}.
\item An informative error message is thrown and the execution is stopped if:
\itemize{
\item \code{x} is not a numeric vector of valid probability values.
\item \code{x} is a numeric vector of valid probability values with some \code{NA} or \code{NaN} values and \code{allow_nas} is set to \code{FALSE}.
}
}
}
\description{
\code{inspect_prob} checks if an object is a numeric vector of valid probability values. This can be useful to validate inputs, intermediate calculations or outputs in user-defined functions.
}
\details{
\code{inspect_prob} conducts a series of tests to check if \code{x} is a numeric vector of valid probability values. Namely, \code{inspect_prob} checks if:
\itemize{
\item \code{x} is \code{NULL} or empty.
\item \code{x} is a numeric, atomic vector.
\item \code{x} has \code{NA} or \code{NaN} values.
\item The values of \code{x} are in the [0, 1] interval.
}
}
\examples{
# Calls that pass silently:
x <- c(0.1, 0.2, 0.3, 0.4, 0.5)
inspect_prob(x)
inspect_prob(seq(0, 1, 0.1))

# Calls that throw an informative warning message:
y <- c(0.1, 0.2, NA, 0.4, 0.5)
\dontrun{inspect_prob(y)}
\dontrun{inspect_prob(y, allow_nas = TRUE)}

# Calls that throw an informative error message:
z1 <- c(-0.9, 0, 0.1, 0.2, 0.3, 0.4, 0.5)
\dontrun{inspect_prob(z1)}
z2 <- c(-0.9, 0, 0.1, 0.2, 0.3, 0.4, 0.5)
\dontrun{inspect_prob(z2, allow_nas = FALSE)}
mylist <- list(NULL, TRUE, factor(.5), matrix(0.5),
         "0.5", list(0.5), NA, NaN, 1.1, .5)
\dontrun{inspect_prob(mylist)}
\dontrun{inspect_prob(mylist[[1]])}
\dontrun{inspect_prob(mylist[[2]])}
\dontrun{inspect_prob(mylist[[3]])}
\dontrun{inspect_prob(mylist[[4]])}
\dontrun{inspect_prob(mylist[[5]])}
\dontrun{inspect_prob(mylist[[6]])}
\dontrun{inspect_prob(mylist[[7]])}
\dontrun{inspect_prob(mylist[[8]])}
\dontrun{inspect_prob(mylist[[9]])}
\dontrun{inspect_prob(mylist[[10]])}

}
\seealso{
\itemize{
\item \code{\link[inspect]{inspect_bf}} to check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[inspect]{inspect_log_bf}} to check if an object is a numeric vector of valid logarithmic Bayes factor values.
\item \code{\link[inspect]{inspect_log_base}} to check if an object is a numeric vector of \code{\link[base]{length}} 1 representing a valid logarithmic base.
\item \code{\link[inspect]{inspect_scale}} to check if an object is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package.
}
}
