% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_true_or_false}
\alias{inspect_true_or_false}
\title{Check if an object is TRUE or FALSE}
\usage{
inspect_true_or_false(x)
}
\arguments{
\item{x}{An arbitrary object.}
}
\value{
\code{inspect_scale} does not return any output. There are two possible scenarios:
\itemize{
\item The call is silent if \code{scale} is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package.
\item An informative error message is thrown otherwise.
}
}
\description{
\code{inspect_true_or_false} checks if an object is a logical vector of \code{\link[base]{length}} with the value \code{TRUE} \code{FALSE}. This can be useful to validate inputs in user-defined functions.
}
\details{
\code{inspect_scale} conducts a series of tests to check if \code{scale} is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package. Namely, \code{inspect_scale} checks if:
\itemize{
\item \code{scale} is \code{NULL} or empty.
\item \code{scale} is an atomic vector of type character and \code{\link[base]{length}} 1 specifying either "Jeffreys" or "Kass-Raftery" (not case sensitive).
\item \code{scale} is \code{NA} or \code{NaN}.
}
}
\examples{
# Calls that pass silently:
inspect_true_or_false(TRUE)
inspect_true_or_false(FALSE)
x <- TRUE
inspect_true_or_false(x)

# Calls that throw informative error messages:
\dontrun{mylist <- list(NULL, NA, NaN, 1, 0, "TRUE")}
\dontrun{inspect_true_or_false(mylist)}
\dontrun{inspect_true_or_false(mylist[[1]])}
\dontrun{inspect_true_or_false(mylist[[2]])}
\dontrun{inspect_true_or_false(mylist[[3]])}
\dontrun{inspect_true_or_false(mylist[[4]])}
\dontrun{inspect_true_or_false(mylist[[5]])}
\dontrun{inspect_true_or_false(mylist[[6]])}

}
\seealso{
\itemize{
\item \code{\link[inspect]{inspect_prob}} to check if an object is a numeric vector of valid probability values.
\item \code{\link[inspect]{inspect_bf}} to check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[inspect]{inspect_log_bf}} to check if an object is a numeric vector of valid logarithmic Bayes factor values.
\item \code{\link[inspect]{inspect_log_base}} to check if an object is a numeric vector of \code{\link[base]{length}} 1 representing a valid logarithmic base.
}
}
