% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\docType{methods}
\name{labeledSF}
\alias{labeledSF}
\alias{labeledSF,INSPEcT-method}
\title{Accessor to the slot labeledSF of an INSPEcT object}
\usage{
labeledSF(object)

\S4method{labeledSF}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT}
}
\value{
A numeric that indicates the scaling factors applied between time points
of the data coming from Nascent-seq library (applies directly to synthesis
rates and indirectly to degradation rates)
}
\description{
Accessor to obtain the labeledSF slot associated with the object of class INSPEcT
}
\examples{
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
labeledSF(nascentInspObj10)
}
