% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{libsizes}
\alias{libsizes}
\title{Contains two variables: "nascentLS" and "totalLS" containing the sequencing depth of nascent and total libraries 
respectively, relative to the experiments in "allcounts"}
\format{
numeric vector of length 33
}
\description{
Contains two variables: "nascentLS" and "totalLS" containing the sequencing depth of nascent and total libraries 
respectively, relative to the experiments in "allcounts"
}
