% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logLik-methods.R
\name{logLik}
\alias{logLik}
\alias{logLik,INSPEcT_model-method}
\alias{logLik,INSPEcT-method}
\title{Retrieve results of log likelihood test}
\usage{
logLik(object, ...)

\S4method{logLik}{INSPEcT_model}(object, ...)

\S4method{logLik}{INSPEcT}(object, ...)
}
\arguments{
\item{object}{An object of class INSPEcT or INSPEcT_model}

\item{...}{Additional arguments for the generic}
}
\value{
A matrix of log likelihood test results for all the tested model comparisons
}
\description{
This method is used to retrieve all the log likelihood ratio test results for all pairs tested for all genes.
}
\examples{
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
logLik(nascentInspObj10)
}
