% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/makeGtfFiles-methods.R
\docType{methods}
\name{makeExonsGtfFromDb}
\alias{makeExonsGtfFromDb}
\alias{makeExonsGtfFromDb,TxDb-method}
\alias{makeGtfFromDb}
\alias{makeIntronsGtfFromDb}
\alias{makeIntronsGtfFromDb,TxDb-method}
\title{Make a GTF file of exonic or intronic features from a TxDb object}
\usage{
makeExonsGtfFromDb(object, type, filename)

makeIntronsGtfFromDb(object, type, filename)

\S4method{makeExonsGtfFromDb}{TxDb}(object, type, filename)

\S4method{makeIntronsGtfFromDb}{TxDb}(object, type, filename)
}
\arguments{
\item{object}{An object of class TxDb}

\item{type}{Either "gene" or "tx", see details for further information}

\item{filename}{The name of the GTF file that is created}
}
\value{
None
}
\description{
Make a GTF file of exonic or intronic features from a TxDb object
}
\details{
In case type "gene" is chosen, the union of all exons of transcripts associated to a gene is used.
}
\examples{
require(TxDb.Mmusculus.UCSC.mm9.knownGene)
txdb <- TxDb.Mmusculus.UCSC.mm9.knownGene
isActiveSeq(txdb) <- c(TRUE, rep(FALSE, length(isActiveSeq(txdb)) - 1))
res <- makeExonsGtfFromDb(txdb, 'gene', 'exons.gtf')
isActiveSeq(txdb) <- rep(TRUE, length(isActiveSeq(txdb)))
}

