% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRPKMsFromBams-functions.R
\name{makeRPKMsFromBams}
\alias{makeRPKMsFromBams}
\title{Calculate RPKM and count values on introns and exons from bam/sam files}
\usage{
makeRPKMsFromBams(txdb, paths_foursu = NULL, paths_total, by = c("gene",
  "tx"), countMultiMappingReads = FALSE, allowMultiOverlap = FALSE,
  strandSpecific = 0, isPairedEnd = FALSE, DESeq2 = TRUE,
  temporalDesign = NULL)
}
\arguments{
\item{txdb}{A TranscriptDB object}

\item{paths_foursu}{A vector of paths of 4sU-seq sam files}

\item{paths_total}{A vector of paths of RNA-seq sam files}

\item{by}{A character, either "gene" or "tx", indicating if rpkms and counts should be summarized at the levels of genes or transcripts. "gene" by default}

\item{countMultiMappingReads}{A logical, if multimapping reads should be counted, FALSE by default. Multimap reads are 
identified using the tag "NH" in the bam/sam file.}

\item{allowMultiOverlap}{A logical, indicating if a read is allowed to be assigned to more than one feature, FALSE by default}

\item{strandSpecific}{Numeric, 0 if no strand-specific read counting should be performed, 1 stranded, 2 reversely-stranded. 0 by default}

\item{isPairedEnd}{A logical, if paired-end reads are used, FALSE by default}

\item{DESeq2}{A logical, if TRUE the RPKMs from exons and introns and associated variances are evaluated through the package DESeq2}

\item{temporalDesign}{A numerical which reports the desing of the experiment in terms of time points and replicates. The time points must be ordered according
to the sequence of bam files submitted for the analysis, these labels characterize different files as replicates of a given conditions.}
}
\value{
A list containing rpkms, counts and the annotation extracted from TxDB for exons and introns, if DESeq2 = TRUE the output also contains a set of data
needed to estimate rpkms variances.
}
\description{
Given a TranscriptDb object and a list of bam/sam files for total and eventually RNA 
experiments, "makeRPKMsFromBams" function calculates read counts and RPKM on exonic and intronic 
features per each gene. Reads that fall where intronic and exonic features overlaps are 
univoquely assigned to exons.
}
\examples{
require(TxDb.Mmusculus.UCSC.mm9.knownGene)
txdb <- TxDb.Mmusculus.UCSC.mm9.knownGene

paths_4su <- system.file('extdata', c('bamRep1.bam','bamRep2.bam','bamRep3.bam','bamRep4.bam') , package="INSPEcT")
paths_total <- system.file('extdata', c('bamRep1.bam','bamRep2.bam','bamRep3.bam','bamRep4.bam') , package="INSPEcT")

temporalDesign <- c(0,0,1,1)

#4sU analysis with DESeq2
makeRPKMsOut_4sU <- makeRPKMsFromBams(txdb=txdb,paths_foursu=paths_4su,paths_total=paths_total,temporalDesign=temporalDesign,DESeq2 = TRUE)

rpkms_4sU <- makeRPKMsOut_4sU$rpkms
counts_4sU <- makeRPKMsOut_4sU$counts
annotations_4sU <- makeRPKMsOut_4sU$annotations
dispersion_parameters_DESeq2_4sU <- makeRPKMsOut_4sU$dispersion_parameters_DESeq2

#4sU analysis without DESeq2
makeRPKMsOut_4sU <- makeRPKMsFromBams(txdb=txdb,paths_foursu=paths_4su,paths_total=paths_total,temporalDesign=temporalDesign,DESeq2 = FALSE)

rpkms_4sU <- makeRPKMsOut_4sU$rpkms
counts_4sU <- makeRPKMsOut_4sU$counts
annotations_4sU <- makeRPKMsOut_4sU$annotations

#No4sU analysis with DESeq2
makeRPKMsOut_No4sU <- makeRPKMsFromBams(txdb=txdb,paths_foursu=NULL,paths_total=paths_total,temporalDesign=temporalDesign,DESeq2 = TRUE)

rpkms_No4sU <- makeRPKMsOut_No4sU$rpkms
counts_No4sU <- makeRPKMsOut_No4sU$counts
annotations_No4sU <- makeRPKMsOut_No4sU$annotations
dispersion_parameters_DESeq2_No4sU <- makeRPKMsOut_No4sU$dispersion_parameters_DESeq2
}
