% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{mature}
\alias{mature}
\alias{mature,INSPEcT_steadyNoNascent-method}
\title{Get mature RNA expressions from an object of class INSPEcT_diffsteady}
\usage{
mature(object)

\S4method{mature}{INSPEcT_steadyNoNascent}(object)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}
}
\value{
A matrix containing mature RNA expressions
}
\description{
Extract mature RNA expressions
}
