% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{matureVar}
\alias{matureVar}
\alias{matureVar,INSPEcT_steadyNoNascent-method}
\title{Get mature RNA expressions variances from an object of class INSPEcT_diffsteady}
\usage{
matureVar(object)

\S4method{matureVar}{INSPEcT_steadyNoNascent}(object)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}
}
\value{
A matrix containing mature RNA expressions variances
}
\description{
Extract mature RNA expressions variances
}
