% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{mycerIds10}
\alias{mycerIds10}
\title{An INSPEcT object with evaluated and modeled rates and concentrations}
\format{An INSPEcT object}
\description{
This INSPEcT object contains the evaluated and modeled rates and concentrations
of the very first 10 genes of the dataset rpkms
}
