% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{nascentInspObj}
\alias{nascentInspObj}
\title{An INSPEcT object without modeled rates and concentrations}
\format{An INSPEcT object}
\description{
This INSPEcT object contains the rates first guesses for the entide database in allcounts
}
