% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nCores-methods.R
\docType{methods}
\name{nCores}
\alias{nCores}
\alias{nCores,INSPEcT-method}
\alias{nCores<-}
\alias{nCores<-,INSPEcT,numeric-method}
\title{Get and set number of cores to be used for the modeling}
\usage{
nCores(object, ...)

nCores(object) <- value

\S4method{nCores}{INSPEcT}(object, ...)

\S4method{nCores}{INSPEcT,numeric}(object) <- value
}
\arguments{
\item{object}{An object of class INSPEcT}

\item{...}{Additional arguments for the generic}

\item{value}{A numeric value with the number of cores to be used}
}
\value{
Number of cores to be used
}
\description{
Get and set number of cores to be used for the modeling

This method is a fast way to get and set the number of cores that the modeling part will 
use (method \code{\link{modelRates}}). The same task can be achieved using the more general
\code{\link{modelingParams}} method
}
\examples{
data('mycerIds10', package='INSPEcT')
nCores(mycerIds10)
## use the more general modelingParams method
modelingParams(mycerIds10)$nCores
data('mycerIds10', package='INSPEcT')
mycerIds10 <- removeModel(mycerIds10)
nCores(mycerIds10) <- detectCores()
## use the more general modelingParams method
modelingParams(mycerIds10)$nCores <- detectCores()
}

