% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\docType{methods}
\name{nTpts}
\alias{nTpts}
\alias{nTpts,INSPEcT-method}
\title{Get the number of time points within the INSPEcT object}
\usage{
nTpts(object)

\S4method{nTpts}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT}
}
\value{
A numeric that indicates the number of time points contained the object
}
\description{
Get the number of time points within the INSPEcT object

A method to obtain the number of the tpts associated with the object of class INSPEcT
}
\examples{
data('mycerIds10')
nTpts(mycerIds10)
}
