% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{plotPMgene}
\alias{plotPMgene}
\alias{plotPMgene,INSPEcT_steadyNoNascent-method}
\title{Plot the premature/mature expression of a gene and the global trend from an object of class INSPEcT_diffsteady}
\usage{
plotPMgene(object, gene_id, samples_colors = 1)

\S4method{plotPMgene}{INSPEcT_steadyNoNascent}(object, gene_id, samples_colors = 1)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}

\item{gene_id}{A numeric that indicated the index of the gene to be plotted}

\item{samples_colors}{The color code relative to the samples}
}
\description{
Plot the premature and mature expressions of a specific gene in the 
different samples of the dataset along with the null model and the 
log2 fold change threshold. Individal observations that fall outside of the
dashed lines are considered post-transcriptional events.
}
