% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{plotPMtrend}
\alias{plotPMtrend}
\alias{plotPMtrend,INSPEcT_steadyNoNascent-method}
\title{Plot the premature/mature trend from an object of class INSPEcT_diffsteady}
\usage{
plotPMtrend(inspectIds)

\S4method{plotPMtrend}{INSPEcT_steadyNoNascent}(inspectIds)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}
}
\description{
Plot the null model estimated for the specific dataset. The null model
is the trend between premature and mature expression, which is usually
linear in the log-log scale and generally points to an increase in the
ratio between premature and mature RNA at increased levels of 
expression
}
