% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{premature}
\alias{premature}
\alias{premature,INSPEcT_steadyNoNascent-method}
\title{Get premature RNA expressions from an object of class INSPEcT_diffsteady}
\usage{
premature(object)

\S4method{premature}{INSPEcT_steadyNoNascent}(object)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}
}
\value{
A matrix containing premature RNA expressions
}
\description{
Extract premature RNA expressions
}
