% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{prematureVar}
\alias{prematureVar}
\alias{prematureVar,INSPEcT_steadyNoNascent-method}
\title{Get premature RNA expressions variances from an object of class INSPEcT_diffsteady}
\usage{
prematureVar(object)

\S4method{prematureVar}{INSPEcT_steadyNoNascent}(object)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}
}
\value{
A matrix containing premature RNA expressions variances
}
\description{
Extract premature RNA expressions variances
}
