% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{PTratio}
\alias{PTratio}
\alias{PTratio,INSPEcT_steadyNoNascent-method}
\title{Calculate post-transcriptional ratio from an object of class INSPEcT_diffsteady}
\usage{
PTratio(object, infToNA = TRUE)

\S4method{PTratio}{INSPEcT_steadyNoNascent}(object, infToNA = TRUE)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}

\item{infToNA}{A logical indicating whether infinite values (originating from zero 
valued premature expressions) should be set artificially to NA or not}
}
\value{
A matrix containing the PTratios
}
\description{
Extract the ratio between mature and premature RNAs
}
