% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/steadyNoNascent-methods.R
\name{PTreg}
\alias{PTreg}
\alias{PTreg,INSPEcT_steadyNoNascent-method}
\title{Calculate the post-transcriptional ratio from an object of class INSPEcT_diffsteady}
\usage{
PTreg(object)

\S4method{PTreg}{INSPEcT_steadyNoNascent}(object)
}
\arguments{
\item{object}{An object of class INSPEcT_steadyNoNascent}
}
\value{
A matrix containing the post-transcriptional regulated genes.
This matrix is generated by the method compareSteadyStateNoNascent. 
It generally report 1 for regulated genes in specific samples, 0 for 
non regulated genes and NA for genes that do not passed the 
expression threshold. In case the argument returnNormScores was set 
to TRUE, instead of discretes values, the deviations from the 
expected model normalized by the experimental standard deviation
is reported.
}
\description{
Extract the post-transcriptional regulation matrix
}
