% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyExpressionsFromTrAbundance.R
\name{quantifyExpressionsFromTrAbundance}
\alias{quantifyExpressionsFromTrAbundance}
\title{Given introns and exons abundances (for example RPKMs) this method returns their variances evaluated thorugh plgem.}
\usage{
quantifyExpressionsFromTrAbundance(exonsAbundances, intronsAbundances,
  experimentalDesign, varSamplingCondition, simulatedData = FALSE)
}
\arguments{
\item{exonsAbundances}{A matrix containing the exons abundances for each experimental condition.}

\item{intronsAbundances}{A matrix containing the intorns abundances for each experimental condition.}

\item{experimentalDesign}{A numerical which reports the desing of the experiment in terms of time points and replicates. The time points must be ordered according
to the columns of the count matrices submitted for the analysis; these labels define conditions and replicates.}

\item{varSamplingCondition}{A character reporting which experimental condition should be used to sample the variance if DESeq2 = FALSE.}

\item{simulatedData}{A boolean which is TRUE if the data under analysis are simulated.}
}
\value{
A list containing RPKMs and associated variances for exons and introns.
}
\description{
Given introns and exons abundances (for example RPKMs) this method returns their variances evaluated thorugh plgem.
}
