% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ratePvals-methods.R
\name{ratePvals}
\alias{ratePvals}
\alias{ratePvals,INSPEcT-method}
\title{Retrieve a single p-value for each rate}
\usage{
ratePvals(object)

\S4method{ratePvals}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT}
}
\description{
This method is used to retrieve all the p-values relative to the variability of synthesis, processing and degradation rates.
}
\examples{
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
ratePvals(nascentInspObj10)
}
