% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{rpkms}
\alias{rpkms}
\title{intron, exons RPKM values of 500 genes from 4sU-seq and RNA-seq experiments}
\format{A list of 4 matrices with 500 rows and 9 columns}
\description{
A dataset containing the values of exonic and intronic RPKMs of 500 genes
both in 4sU-seq and RNA-seq experiments. The variables are as follows:
}
\details{
\itemize{
  \item rpkms_4su_exons
  \item rpkms_4su_introns
  \item rpkms_total_exons
  \item rpkms_total_introns
}
}

