% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/setConfidenceIntervals-method.R
\name{setConfidenceIntervals}
\alias{setConfidenceIntervals}
\alias{setConfidenceIntervals,INSPEcT-method}
\title{Set confidence intervals}
\usage{
setConfidenceIntervals(object, confidenceIntervals)

\S4method{setConfidenceIntervals}{INSPEcT}(object, confidenceIntervals)
}
\arguments{
\item{object}{An object of class INSPEcT_model}

\item{confidenceIntervals}{list of confidence intervals.}
}
\value{
An object of class ExpressionSet containing the confidence intervals.
}
\description{
This function is used to set the confidence intervals in the nascent RNA mode.
}
