% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{simRates}
\alias{simRates}
\title{An INSPEcT_model object of 1000 simulated rates}
\format{An INSPEcT_model object}
\description{
The INSPEcT_model object contains 1000 simulated rates that were obtained
using the dataset rpkms as reference
}

