% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\docType{methods}
\name{totalSF}
\alias{totalSF}
\alias{totalSF,INSPEcT-method}
\title{Accessor to the slot totalSF of an INSPEcT object}
\usage{
totalSF(object)

\S4method{totalSF}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT}
}
\value{
A numeric that indicates the scaling factors applied between time points
of the data coming from RNA-seq library (applies to total-mRNAs
and pre-mRNAs concentrations)
}
\description{
Accessor to the slot totalSF of an INSPEcT object

Accessor to obtain the totalSF slot associated with the object of class INSPEcT
}
\examples{
data('mycerIds10')
totalSF(mycerIds10)
}

