% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{tpts}
\alias{tpts}
\alias{tpts,INSPEcT-method}
\title{Accessor to the slot tpts of an INSPEcT object}
\usage{
tpts(object)

\S4method{tpts}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT}
}
\value{
A numeric that indicates time points contained the object
}
\description{
Accessor to obtain the tpts associated with the object of class INSPEcT
}
\examples{
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
tpts(nascentInspObj10)
}
