% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{UCSC_mm9_genes_exons}
\alias{UCSC_mm9_genes_exons}
\title{A Gtf file containing exons definition of 100 genes}
\format{A tab separated file}
\description{
A Gtf file containing exons definition of 100 genes
}

