% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\name{UCSC_mm9_genes_introns}
\alias{UCSC_mm9_genes_introns}
\title{A Gtf file containing introns definition of 100 genes}
\format{A tab separated file}
\description{
A Gtf file containing introns definition of 100 genes
}

