% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/compareSteadyNoNascent-methods.R
\docType{methods}
\name{compareSteadyNoNascent}
\alias{compareSteadyNoNascent}
\alias{compareSteadyNoNascent,INSPEcT-method}
\title{Compare mature RNA steady state data from an object of class INSPEcT}
\usage{
compareSteadyNoNascent(inspectIds, expressionThreshold = 0.25,
  log2FCThreshold = 2)

\S4method{compareSteadyNoNascent}{INSPEcT}(inspectIds,
  expressionThreshold = 0.25, log2FCThreshold = 2)
}
\arguments{
\item{inspectIds}{An object of class INSPEcT.}

\item{expressionThreshold}{A parameter which sets how many log2 fold changes of distance from the median behaviour are imputable to noise.}

\item{log2FCThreshold}{A parameter which sets the log2 fold change distance from the median behaviour that is imputable to noise.}
}
\description{
Compare mature RNA steady state data from an object of class INSPEcT

This function compare exons and introns expression level matrices, from two up to an arbitrary number of samples, in order
to identify genes which are oddly regluated, compared to an expected standard behaviour, from the post transcriptional point of view.
}
\examples{
data('allcounts', package='INSPEcT')
data('featureWidths', package='INSPEcT')
data('libsizes', package='INSPEcT')

nascentCounts<-allcounts$nascent
matureCounts<-allcounts$mature
conditions<-letters[1:11]
expDes<-rep(conditions,3)

matExp_DESeq2<-quantifyExpressionsFromTrCounts(
      allcounts=matureCounts
      ,libsize=totalLS
      ,exonsWidths=exWdths
      ,intronsWidths=intWdths
      ,experimentalDesign=expDes)

matureInspObj <- newINSPEcT(tpts=conditions,matureExpressions=matExp_DESeq2)

regGenes<-compareSteadyNoNascent(inspectIds=matureInspObj
								   ,expressionThreshold=0.25
							   ,log2FCThreshold=2.)
head(regGenes)
table(regGenes)
}
