% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/computeConfidenceIntervals.R
\name{computeConfidenceIntervals}
\alias{computeConfidenceIntervals}
\alias{computeConfidenceIntervals,INSPEcT-method}
\title{Compute confidence intervals}
\usage{
computeConfidenceIntervals(object, BPPARAM = bpparam())

\S4method{computeConfidenceIntervals}{INSPEcT}(object, BPPARAM = bpparam())
}
\arguments{
\item{object}{An object of class INSPEcT_model}

\item{BPPARAM}{Parallelization parameters for bplapply. By default bpparam()}
}
\value{
An object of class INSPEcT.
}
\description{
This function is used to compute the confidence intervals for a given set of modeled genes in the NoNascent scenario.
}
