% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{package}
\name{INSPEcT}
\alias{INSPEcT}
\alias{INSPEcT-package}
\title{INSPEcT package}
\description{
INSPEcT (INference of Synthesis, Processing and dEgradation rates in Time-cousre experiments) 
is a package that analyse 4sU-seq and RNA-seq time-course data in order to evaluate synthesis, 
processing and degradation rates and asses via modeling the rates that determines changes in mature mRNA levels.

It implements two classes (\code{\linkS4class{INSPEcT_model}} and 
\code{\linkS4class{INSPEcT}}) and their corresponding methods. To have a detailed 
description of how the two classes are structured and which methods apply on, type:

\code{?'INSPEcT-class'} \cr
\code{?'INSPEcT_model-class'} \cr
\code{?'INSPEcT_diffsteady-class'}

To see how the typical workflow of INSPEcT works, type: \cr
\code{vignette('INSPEcT')}

Last but not least, to obtain a citation, type: \cr
\code{citation('INSPEcT')}
}

