% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_categories}
\alias{inspect_categories}
\title{lorem ipsum}
\usage{
inspect_categories(x)
}
\arguments{
\item{x}{lorem ipsumlorem ipsum}
}
\value{
lorem ipsum
}
\description{
lorem ipsum
}
\details{
lorem ipsum
}
