% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_data_.R
\name{inspect_data_multinomial}
\alias{inspect_data_multinomial}
\title{Validate Categorical data}
\usage{
inspect_data_multinomial(data, warning_nas = FALSE)
}
\arguments{
\item{data}{An arbitrary object.}

\item{warning_nas}{Logical vector of \code{\link[base]{length}} 1. If \code{TRUE} then the presence of \code{NA} or \code{NaN} values in \code{data} generates a warning.}
}
\value{
\code{inspect_data_multinomial} does not return any output. There are three possible outcomes:
\itemize{
\item The call is silent if:
\itemize{
\item \code{data} is eligible to have been generated by a Multinomial distribution and there are no \code{NA} or \code{NaN} values in \code{data}
\item \code{data} is eligible to have been generated by a Multinomial distribution, there are some \code{NA} or \code{NaN} values in \code{data} and \code{warning_nas} is set to \code{FALSE}.
}
\item An informative warning message is thrown if \code{data} is eligible to have been generated by a Multinomial distribution, there are some \code{NA} or \code{NaN} values in \code{data} and \code{warning_nas} is set to \code{TRUE}.
\item An informative error message is thrown and the execution is stopped if \code{data} is not eligible to have been generated by a Multinomial distribution.
}
}
\description{
\code{inspect_data_multinomial} checks if an object contains data that is eligible to have been generated by a Multinomial distribution. This can be useful to validate inputs in user-defined functions.
}
\details{
\code{inspect_data_multinomial} conducts a series of tests to check if \code{data} is eligible to have been generated by a Multinomial distribution. Namely, \code{inspect_data_multinomial} checks if:
\itemize{
\item \code{data} is \code{NULL} or empty.
\item \code{data} is atomic and have an eligible data type (logical, integer, double, character).
\item \code{data} has \code{NA} or \code{NaN} values.
}
}
\examples{
# Calls that pass silently:
x1 <- c(1, 0, 0, 1, 2)
x2 <- c(FALSE, FALSE, TRUE, NA)
x3 <- c("yes", "no", "yes", "maybe")
x4 <- factor(c("yes", "no", "yes", "maybe"))
x5 <- c(1, 0, 0, 1, 0, NA, 2)
inspect_data_multinomial(x1)
inspect_data_multinomial(x2)
inspect_data_multinomial(x3)
inspect_data_multinomial(x4)
inspect_data_multinomial(x5)
inspect_data_multinomial(x5)

# Call that throws an informative warning message:
\dontrun{y1 <- c(1, 1, NA, 0, 0, 2)}
\dontrun{inspect_data_multinomial(y1, warning_nas = TRUE)}

# Calls that throw an informative error message:
\dontrun{inspect_data_multinomial(NULL)}
\dontrun{inspect_data_multinomial(list(1,0))}
\dontrun{inspect_data_multinomial(numeric(0))}
\dontrun{inspect_data_multinomial(NaN)}
\dontrun{inspect_data_multinomial(NA)}

}
