% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_.R
\name{inspect_log_base}
\alias{inspect_log_base}
\title{Check if an object is a valid logarithmic base}
\usage{
inspect_log_base(base)
}
\arguments{
\item{x}{An arbitrary object.}
}
\value{
\code{inspect_log_base} does not return any output. There are two possible outcomes:
\itemize{
\item The call is silent if \code{x} is a numeric vector of \code{\link[base]{length}} 1 that is a valid logarithmic base.
\item An informative error message is thrown otherwise.
}
}
\description{
\code{inspect_log_base} checks if an object is a numeric vector of \code{\link[base]{length}} 1 that is eligible to be used as a logarithmic base. This can be useful to validate inputs in user-defined functions.
}
\details{
\code{inspect_log_base} conducts a series of tests to check if \code{x} is a valid logarithmic base. Namely, \code{inspect_log_base} checks if:
\itemize{
\item \code{x} is \code{NULL} or empty.
\item \code{x} is a numeric, atomic vector of \code{\link[base]{length}} 1.
\item \code{x} is \code{NA} or \code{NaN}.
}
}
\examples{
# Calls that pass silently:
inspect_log_base(10)
inspect_log_base(2)
inspect_log_base(exp(1))
inspect_log_base(0.5)

# Calls that throw informative error messages:
\dontrun{inspect_log_base(NULL)}
\dontrun{inspect_log_base(numeric(0))}
\dontrun{inspect_log_base(-1)}
\dontrun{inspect_log_base(0)}
\dontrun{inspect_log_base(TRUE)}
\dontrun{inspect_log_base("10")}
\dontrun{inspect_log_base(factor(10))}
\dontrun{inspect_log_base(matrix(10))}
\dontrun{inspect_log_base(list(10))}
\dontrun{inspect_log_base(NA)}
\dontrun{inspect_log_base(NaN)}

}
\seealso{
\itemize{
\item \code{\link[inspect]{inspect_prob}} to check if an object is a numeric vector of valid probability values.
\item \code{\link[inspect]{inspect_bf}} to check if an object is a numeric vector of valid Bayes factor values.
\item \code{\link[inspect]{inspect_log_bf}} to check if an object is a numeric vector of valid logarithmic Bayes factor values.
\item \code{\link[inspect]{inspect_scale}} to check if an object is a string of characters representing one of the Bayes factor interpretation scales available in the \code{pcal} package.
}
}
