% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector.R
\docType{package}
\name{inspector-package}
\alias{inspector}
\alias{inspector-package}
\title{The \{inspector\} Package}
\description{
Validation of arguments and objects in user-defined functions
}
\details{
The \code{inspector} package provides utility functions that implement and automate common sets of validation tasks. These functions are particularly useful to validate inputs, intermediate objects and output values in user-defined functions, resulting in tidier and less verbose functions.

For a complete list of functions see \code{library(help = "inspector")}. For instructions and examples see the "Usage" section of the \href{https://github.com/pedro-teles-fonseca/inspector/blob/master/README.md}{README}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedro-teles-fonseca.github.io/inspector/}
  \item \url{https://github.com/pedro-teles-fonseca/inspector}
  \item Report bugs at \url{https://github.com/pedro-teles-fonseca/inspector/issues}
}

}
\author{
\strong{Maintainer}: Pedro Fonseca \email{pedro.teles.fonseca@phd.iseg.ulisboa.pt} (\href{https://orcid.org/0000-0002-5708-081X}{ORCID})

Other contributors:
\itemize{
  \item FCT [funder]
}

}
\keyword{internal}
