% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeRPKMs-functions.R
\name{makeRPKMs}
\alias{makeRPKMs}
\title{Calculate RPKM values on introns and exons from sam files}
\usage{
makeRPKMs(tpts, sampaths_4su, sampaths_total, exons_gtf_file, introns_gtf_file,
  out_folder = ".")
}
\arguments{
\item{tpts}{A vector of time points corresponding in which experiments have been collected (it admits non-unique values in case of replicates)}

\item{sampaths_4su}{A vector of paths of 4sU-seq sam files of the same length of tpts}

\item{sampaths_total}{A vector of paths of RNA-seq sam files of the same length of tpts}

\item{exons_gtf_file}{A path to the GTF file that defines exon features}

\item{introns_gtf_file}{A path to the GTF file that defines intron features (if NULL only exons reads are quantified)}

\item{out_folder}{A path to the output folder (by default the current directory)}
}
\value{
A list of intron and exon rpkms for both RNA-seq and 4sU-seq
}
\description{
Given a vector of time points, paths to sam files of the 4sU-seq and RNA-seq experiments and paths of GTF files, "makeRPKMs" function calculates read counts and RPKM on exonic and intronic features per each gene. Reads that fall where intronic and exonic features overlaps are assigned to exons.
This function is based on HTSeq software (http://www-huber.embl.de/users/anders/HTSeq/doc/overview.html) which is expected to pre installed and on the path.
}
\examples{
exons_gtf_file <- system.file('extdata', 'UCSC_mm9_genes_exons.txt', package="INSPEcT")
introns_gtf_file <- system.file('extdata', 'UCSC_mm9_genes_introns.txt', package="INSPEcT")
samfiles4su <- system.file('extdata', '4sURNA_0h.sam', package="INSPEcT")
samfilesTotal <- system.file('extdata', 'totalRNA_0h.sam', package="INSPEcT")
\dontrun{
rpkms <- makeRPKMs(1, samfiles4su, samfilesTotal, exons_gtf_file, introns_gtf_file)
save(rpkms, file='rpkms.rd')
}
}

