% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/show-methods.R
\name{nGenes}
\alias{nGenes}
\alias{nGenes,INSPEcT-method}
\title{Get the number of genes within the INSPEcT object}
\usage{
nGenes(object)

\S4method{nGenes}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT}
}
\value{
A numeric that indicates the number of genes within the object
}
\description{
A method to obtain the number of the genes associated with the object of class INSPEcT
}
\examples{
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
nGenes(nascentInspObj10)
}
