% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady-state-methods.R
\docType{methods}
\name{plotMA}
\alias{plotMA}
\alias{plotMA,INSPEcT_diffsteady-method}
\title{MA-plot from base means and log fold changes}
\usage{
\S4method{plotMA}{INSPEcT_diffsteady}(object, ...)
}
\arguments{
\item{object}{An object of calss INSPEcT_diffsteady}

\item{...}{Additional parameters, see Details section}
}
\description{
Visualize the comparison between the rates calculated from two different INSPEcT objects
profiled in steady-state conditions.
}
\details{
Possible arguments to "plotMA":
\itemize{
\item "rate" - A character, which represent the rate to be visualized, either "synthesis", "processing" or "degradation". By default, "synthesis" is chosen.
\item "alpha" - A numeric, The confidence interval for significance (FDR), by default 0.1
\item "xlim" - A numeric vector of length 2, limits of x-axis, by default the range of the data.
\item "xlab" - A character, the label of x-axis, by default "log2 geometric mean"
\item "ylim" - A numeric vector of length 2, limits of y-axis, by default the range of the data.
\item "ylab" - A character, the label of y-axis, by default "log2 fold change"
\item "main" - A character, the title of the plot, by default the name of the visualized rate.
}
}
\examples{
data('nascentInspObj', package='INSPEcT')	
simRates<-makeSimModel(nascentInspObj, 1000, seed=1)
newTpts<-c(0, 16)
nascentInspObj_sim3<-makeSimDataset(object=simRates
                                   ,tpts=newTpts
                                   ,nRep=3
                                   ,NoNascent=FALSE
                                   ,seed=1)
newTpts<-c(0, 16)
nascentInspObj_sim4<-makeSimDataset(object=simRates
                                   ,tpts=newTpts
                                   ,nRep=4
                                   ,NoNascent=FALSE
                                   ,seed=2)

diffrates<-compareSteady(inspectIds1=nascentInspObj_sim3
                        ,inspectIds2=nascentInspObj_sim4
                        ,nRep1=3
                        ,nRep2=4)
plotMA(diffrates, alpha=.95)
}
\seealso{
\url{http://en.wikipedia.org/wiki/MA_plot}
}
