% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sfPlot-methods.R
\docType{methods}
\name{sfPlot}
\alias{sfPlot}
\alias{sfPlot,INSPEcT-method}
\title{A nice plot to see scaling factors used for RNA-seq and Nascent-seq libraries}
\usage{
sfPlot(object)

\S4method{sfPlot}{INSPEcT}(object)
}
\arguments{
\item{object}{An object of class INSPEcT}
}
\value{
None
nascentInspObj10 <- readRDS(system.file(package='INSPEcT', 'nascentInspObj10.rds'))
sfPlot(nascentInspObj10)
}
\description{
This method generates a plot that immediately shows the scaling factors used to scale
RNA- and Nascent-seq libraries and the possible relations between them. The ratio between the RNA- and the 
Nascent-seq scaling can be in fact considered as a yield of the synthesis within the cells.
}
